reorganize project structure and add READMEs (#21)
* reorganize project structure and add READMEs

* E5CN flowcharts

* initial work on CathodeHeatingSubsystem flowcharts

* power supply initialization flowchart

* added readme and driver file for G9SP

* added the Readme for the interlocks.py

* made changes to the readme

* corrected spacings on README.md

* missing png file

* updated GUI layout image

* rewrite top-level readme

* use relative filepath for GUI layout graphic

* more descriptive folder name for G9

* initial work on 9104 README

* basic usage for E5CN README

* clean up usage

* hardware specification readme updates

* restructure README.md

* working on flow chart for g9driver

* small changes flow chart for the process output

* updated interlocks readme to have some more info related to some expected values

---------
