+++ instrumentctl/__init__.py
@@ -1,8 +1,11 @@
-from .apex_mass_flow_controller import ApexMassFlowController
-from .power_supply_9104 import PowerSupply9104
+from .apex_mass_flow.apex_mass_flow_controller import ApexMassFlowController
+from .power_supply_9104.power_supply_9104 import PowerSupply9104
+from .E5CN_modbus.E5CN_modbus import E5CNModbus
+from .G9SP_interlock.g9_driver import G9Driver
 
 __all__ = [
     'ApexMassFlowController',
     'PowerSupply9104',
-    'E5CNModbus'
+    'E5CNModbus',
+    'G9Driver',
 ]
