+++ subsystem/__init__.py
@@ -1,9 +1,12 @@
-from .vtrx import VTRXSubsystem
-from .environmental import EnvironmentalSubsystem
-from .interlocks import InterlocksSubsystem
-from .oil_system import OilSubsystem
-from .cathode_heating import CathodeHeatingSubsystem
-from .visualization_gas_control import VisualizationGasControlSubsystem
+from .vtrx.vtrx import VTRXSubsystem
+from .environmental.environmental import EnvironmentalSubsystem
+from .interlocks.interlocks import InterlocksSubsystem
+from .oil_system.oil_system import OilSubsystem
+from .cathode_heating.cathode_heating import CathodeHeatingSubsystem
+from .visualization_gas_control.visualization_gas_control import VisualizationGasControlSubsystem
+from .beam_extraction.beam_extraction import BeamExtractionSubsystem
+from .beam_pulse.beam_pulse import BeamPulseSubsystem
+from .deflection_monitor.deflection_monitor import DeflectionMonitorSubsystem
 
 __all__ = [
     'VTRXSubsystem',
@@ -11,5 +14,8 @@ __all__ = [
     'InterlocksSubsystem',
     'OilSubsystem',
     'CathodeHeatingSubsystem',
-    'VisualizationGasControlSubsystem'
+    'VisualizationGasControlSubsystem',
+    'BeamExtractionSubsystem',
+    'BeamPulseSubsystem',
+    'DeflectionMonitorSubsystem'
 ]
