+++ subsystem/beam_extraction/beam_extraction.py
@@ -1,0 +1,9 @@
+
+class BeamExtractionSubsystem:
+    """
+    Manages the beam extraction system
+    """
+
+    def __init__(self):
+        # TODO: Implement
+        pass
